*cleaned all tmpl-files for tabs and "bad html", which generally just means doubleqouting instead of single-qouting or no-qouting.
*replaced the code for the title. was <h1 class="headline">.
now the title is built into a table together with the
breadcrumb, thus making it easier to adjust and works cross-browser.


*"fixed" frontpage to contain invisble space so titlebar of
frontpage is in the exact same position as all others for
nice browing pleasure  [:-)] . maybe we could have some graphic
above the "welcome ..." title, it seems empty without the
breadcrumb, but i prefer empiness before a "dancing" titlebar.


*"fixed" the two next.gifs to be 48 wide indtead of 49 which
caused ns4 to show the outer right-frame of those pics


*removed the <div content> completely as it does no good, i
think. maybe i have overlooked something here so we can just
insert it again, but i dont see the point. we are now having
aligned margins for actual text-content, breadcrumb, search,
navs etc.


*the menu is now looking the same (more or less) across
"all" browsers. it is done with an accidently discovered
hack and without the use of space-gifs. it now has a
seperator between menuitems, but this can be easily replaced
if we dont want these seperators. i have added the
medium-blue color in the titles for the menus and the dark
blue as background for selected menu item. im not sure which
combination of colours work the best but i liked these boxes
titles to have a color that was not gray.


*the ad section now also looks the same in all browsers,
done with the same trick as for menus. book-title same as
above. i also added black 1px border around the images which
i think is a classic mistake not to have around any ads of
this sort  [:)]
NB! i have overlooked the bg-color for these ads (they
should have the same light gray colour as the menu-boxes, i
have a fix made locally)


*in the download section i have removed the <div> and put
the class-selector into the tablecell alredy there. i have
removed the parenthesis and made the file-size part of the
link, which is very usual for downloads, so now that text is
also coloured like other links. i have reduced the sizes of
the images, mainly just for trying it out (NB! i do not have
a proper pdf-gif so that looks pretty bad) please clear your
caches anyway.

*fonts*
- i have used the helvetica, arial, verdana font-family, and
for certain items i have reversed that order so verdana
shines thru just for spicing it slightly up (breadcrumb, menu)

*the <form>-tags and the hidden fields have been placed
elsewhere otherwise they mess things up for ns4 and also
ie/windows i think. i have stylized the text-field a little
for viewing pleasure in certain browsers (only ie5+ ??)

*i have replaced <hr> w. <hr noshade size="1"> to give it a
clean elegant look


*in the breadcrumb i have replaced space/space, with &nbsp;/&nbsp;


*in navbar_local_bottomi have asigned 50% width for the tow
possible cells


*in navbar_local_top i have removed <div class="download">
and asigned 33% width for the three possible cells


*i have replaced the single &nsb; with a single <br> to
ensure 1px width of the new "pre-bar"


*style.css*
complelety re-arranged and optimzed for ns4 to still be
working with other browsers

- removed all styles that are not used (in this version)
- added visited link colours for all links apart from those
in menu and list-types
- added hover link colours for all links apart from list-types
- added padding for div in ad-areas
- changed .table-top bgcolor to our old medium-blue ASF and
set font color to black!!!
- decreased font-size for navbarglobal to 0.85em
- inserted color for navbarglobal so slashed and links are
now same colours
- increased tail-font-size to 0.85em
- specified right-margin to 0px for ns4
- replaced and ns-optimzed all div.selectmenuitem and
div.notselectmenuitem with td.selectmenuitem and
td.notselectmenuitem respectively
- inserted styles for dl, ul, p
- gave .table-top an absolute font-size like other menu-items

PR:
Obtained from:
Submitted by:	allan <lambretta@inet.uni2.dk>
Reviewed by:	stas


git-svn-id: https://svn.apache.org/repos/asf/perl/modperl/docs/trunk@65902 13f79535-47bb-0310-9956-ffa450edef68