The only remaining issue in this todo/release entry regarding <Perl> sections was that you can't write <Perl>1;</Perl> to trigger an early interpreter startup, for instance.
This is a limitation of httpd's configuration parser and is just not
practical to try and fix. Instead, document the absence of this feature,
and suggest using multi-line blocks.

Another worthy possibility for early interpreter startup and perl one-liners
is to use the (currently under-documented) Perl directive in httpd.conf :

  Perl 1;
  Perl warn "Hello World";

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/perl/modperl/docs/trunk@67754 13f79535-47bb-0310-9956-ffa450edef68